A JS project was created. The file arifmetics.js was formed, in which at least two variables of different types were added, and arithmetic operations were performed on both same-type and different-type variables. The results of the operations were output to the console.
The file logical.js was created, in which at least one variable of different types was added, and logical operations (comparison operations and logical operators) were performed. The results of the operations were output to the console.

The file decision-tree.js was prepared, in which a branching structure using if and if-else if-else with comparison operations and logical operators was written.

The file switch.js was created, in which the if-else if-else branching structure from decision-tree.js was rewritten using switch.